硬件解码

+++ SnailFFmpeg/app/src/main/java/com/snail/snailffmpeg/base/IDecoder.kt
@@ -27,7 +27,7 @@ interface IDecoder : Runnable {
 
     fun getWidth(): Int
 
-    fun getDuration(): Int
+    fun getDuration(): Long
 
     fun getRotationAngle(): Int
